requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
PyMuPDF==1.23.3
requests==2.31.0
pydantic==2.4.2
numpy>=1.23.0,<2.0.0
python-jose==3.3.0
passlib==1.7.4
openai-whisper==20231106
torchaudio==2.1.0
TTS==0.21.0
pyttsx3==2.90
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.25.0
aiofiles==23.2.1